Moved the e2e tests into the component directories and updated the jest configs
