+++ packages/core/src/components/HelpDrawer/HelpDrawer.e2e.test.js
@@ -1,10 +1,14 @@
 import 'chromedriver';
-import { ROOT_URL, RULESET_ALL } from '../helpers/constants';
+import { ROOT_URL, RULESET_ALL } from '../../helpers/e2e/constants';
 import WebDriver, { Builder } from 'selenium-webdriver';
+import {
+  getElementByClassName,
+  getElementById,
+  getElementByXPath
+} from '../../helpers/e2e';
 import AxeBuilder from 'axe-webdriverjs';
-import { getElementByClassName } from '../helpers/e2eTestHelpers';
 
-const rootURL = `${ROOT_URL}/example/components.form-label.react/`;
+const rootURL = `${ROOT_URL}/example/components.help-drawer.react-help-drawer/`;
 let driver, el;
 
 beforeEach(() => {
@@ -23,19 +27,35 @@ afterEach(() => {
   driver.quit();
 });
 
-describe('Form Label component', () => {
+describe('Help Drawer component', () => {
   it('Should render', async() => {
     await driver.get(rootURL);
 
-    el = await getElementByClassName(driver, 'ds-c-label');
+    el = await getElementById(driver, 'js-example');
+    expect(el).toBeTruthy();
+  });
+
+  it('Should open the help drawer on click', async() => {
+    await driver.get(rootURL);
+
+    el = await getElementByXPath(driver, '//*[@id="js-example"]/div/span/a');
+    el.click();
+
+    el = await getElementByClassName(driver, 'ds-c-help-drawer');
     expect(el).toBeTruthy();
   });
 
   it('Should have no accessibility violations', async done => {
     await driver.get(rootURL);
 
+    el = await getElementByXPath(driver, '//*[@id="js-example"]/div/span/a');
+    el.click();
+
+    el = await getElementByClassName(driver, 'ds-c-help-drawer');
+
     await AxeBuilder(driver)
       .withTags(RULESET_ALL)
+      .disableRules('bypass')
       .analyze((err, results) => {
         if (err) {
           console.log(err);
