+++ packages/core/src/components/MonthPicker/MonthPicker.e2e.test.js
@@ -1,10 +1,10 @@
 import 'chromedriver';
-import { ROOT_URL, RULESET_ALL } from '../helpers/constants';
+import { ROOT_URL, RULESET_ALL } from '../../helpers/e2e/constants';
 import WebDriver, { Builder } from 'selenium-webdriver';
 import AxeBuilder from 'axe-webdriverjs';
-import { getElementByClassName } from '../helpers/e2eTestHelpers';
+import { getElementByClassName } from '../../helpers/e2e';
 
-const rootURL = `${ROOT_URL}/example/components.autocomplete.react/`;
+const rootURL = `${ROOT_URL}/example/components.month-picker.react/`;
 let driver, el;
 
 beforeEach(() => {
@@ -23,11 +23,11 @@ afterEach(() => {
   driver.quit();
 });
 
-describe('Alert component', () => {
+describe('Month Picker component', () => {
   it('Should render', async() => {
     await driver.get(rootURL);
 
-    el = await getElementByClassName(driver, 'ds-u-clearfix ds-c-autocomplete');
+    el = await getElementByClassName(driver, 'ds-c-month-picker');
     expect(el).toBeTruthy();
   });
