+++ packages/core/src/components/Spinner/Spinner.e2e.test.js
@@ -1,11 +1,10 @@
 import 'chromedriver';
-import { ROOT_URL, RULESET_ALL } from '../helpers/constants';
+import { ROOT_URL, RULESET_ALL } from '../../helpers/e2e/constants';
 import WebDriver, { Builder } from 'selenium-webdriver';
 import AxeBuilder from 'axe-webdriverjs';
-import { getElementByClassName } from '../helpers/e2eTestHelpers';
+import { getElementByClassName } from '../../helpers/e2e';
 
-const rootURL = `${ROOT_URL}/example/components.choice.choicelist/`;
-const inverseURL = `${ROOT_URL}/example/components.choice.inversed/`;
+const rootURL = `${ROOT_URL}/example/components.spinner.react/`;
 let driver, el;
 
 beforeEach(() => {
@@ -24,33 +23,16 @@ afterEach(() => {
   driver.quit();
 });
 
-describe('Choice component', () => {
+describe('Spinner component', () => {
   it('Should render', async() => {
     await driver.get(rootURL);
 
-    el = await getElementByClassName(driver, 'ds-c-fieldset');
+    el = await getElementByClassName(driver, 'ds-c-spinner');
     expect(el).toBeTruthy();
   });
 
   it('Should have no accessibility violations', async done => {
     await driver.get(rootURL);
-
-    await AxeBuilder(driver)
-      .withTags(RULESET_ALL)
-      .analyze((err, results) => {
-        if (err) {
-          console.log(err);
-        }
-        if (results.violations.length >= 1) {
-          console.log(results.violations);
-        }
-        expect(results.violations.length).toBe(0);
-        done();
-      });
-  });
-
-  it('Inverse layout should have no accessibility violations', async done => {
-    await driver.get(inverseURL);
 
     await AxeBuilder(driver)
       .withTags(RULESET_ALL)
