+++ packages/core/src/helpers/e2e/constants.js
