+++ tools/jest/e2e.config.js
@@ -4,7 +4,7 @@ module.exports = {
   globalSetup: '<rootDir>/tools/jest/e2e.global-setup.js',
   globalTeardown: '<rootDir>/tools/jest/e2e.global-teardown.js',
   setupFiles: ['<rootDir>/tools/jest/e2e.polyfills.js'],
-  testMatch: ['<rootDir>/packages/core/src/e2e/**/?*.e2e.test.js'],
+  testMatch: ['<rootDir>/packages/core/src/**/?*.e2e.test.js'],
   transform: {
     '^.+\\.(js|jsx|mjs)$': '<rootDir>/node_modules/babel-jest'
   }
